feat: 词提权重添加词长权重与选重权重

+++ liubility-typing-server/src/main/java/org/liubility/typing/server/code/parse/SubscriptInstance.java
@@ -44,7 +44,7 @@ public class SubscriptInstance {
      * 左右击键差值
      */
     @JsonIgnore
-    private int feelDeviation;
+    private double feelDeviation;
 
 
     @JsonIgnore
